// src/features/trailers/types/trailerApi.ts
export interface TrailerApiResponse<T> {
  status: number;
  message: string;
  data: T;
  hasNext?: boolean;
  page?: number;
  cookies?: Record<string, string>;
}